Exclude active intent UTXOs when creating new transactions
When creating a new transaction intent while another is in review,
exclude the UTXOs that are already being used in the active intent.

Modified getUnconfirmedUTXOs to accept an optional excludeFromIntent
parameter that builds a set of txid:vout keys to exclude from results.

Updated createBtcIntent and createUnitIntent to pass the current
sendIntent to getUnconfirmedUTXOs, preventing UTXO reuse when creating
multiple intents before broadcasting.

This fixes the duplicate input issue when rapidly creating multiple
transaction intents.